Added first round of thermo_clock.py program
Per the readme, this program is based in large part on the program
referenced by Simon Monk on the Adafruit Learning site. I adapted it to
update a thingspeak channel feed and put some special touches for effect
and usability (like the functionality to turn off all blocks of the
display when program is aborted with control-c.